Add WAKE_LOCK permission for download provider.
The larger question is how does the download provider work without it? The download thread acquires a wake lock, so this could never work without WAKE_LOCK permission. Typically it probably (?) gets it from the shared user id android.media shared with MediaProvider, which does request WAKE_LOCK permission. But why don't we? Is it that we just happen not to be loading media provider?